Fixes as requested
